Load dữ liệu, đăng nhập, đăng ký
load dữ liệu sản lên trang index ,  đăng nhập đăng ký tài khoản

+++ src/main/java/com/fpoly/dao/SanPhamDAO.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fpoly.entity.SanPham;
 
-public interface SanPhamDAO extends JpaRepository<SanPham, Integer> {
-	List<SanPham> findAllById(Integer id);
+public interface SanPhamDAO extends JpaRepository<SanPham, Long> {
+	List<SanPham> findAllById(long id);
 }
